added args to sqlalchemy func

+++ main.py
@@ -7,54 +7,61 @@ import os
 import logging
 from file_downloader.companyhouse_transfer import collect_companieshouse_file
 from file_parser.utils import unzip_ch_file, fragment_ch_file, pipeline_messenger, date_check
-from file_parser.fragment_work import parse_fragment, load_fragment
-
+from file_parser.fragment_work import parse_fragment, upsert_sic, upsert_address
+from locker import connect_preprod
 
 logging.basicConfig(level=logging.INFO,
                     format='%(asctime)s [line:%(lineno)d] %(levelname)s: %(message)s')
 logger = logging.getLogger()
 
-host = sys.argv[1]
-user = sys.argv[2]
-passwd = sys.argv[3]
-database = sys.argv[4]
-
+# host = sys.argv[1]
+# user = sys.argv[2]
+# passwd = sys.argv[3]
+# database = sys.argv[4]
+#
+host = 'preprod.cqzf0yke9t3u.eu-west-1.rds.amazonaws.com'
+user = 'rory'
+passwd = 'Me._7;cBsqQ$]JX}'
+database = 'iqblade'
 db = mysql.connector.connect(
     host=host,
     user=user,
     passwd=passwd,
     database=database
 )
-schema = 'iqblade'
 cursor = db.cursor()
+
+schema = 'iqblade'
+
 t_fragment_file = 'BasicCompanyDataAsOneFile-2023-01-01.csv'
 # download file
 logger.info('downloading file')
 firstDayOfMonth = datetime.date(datetime.date.today().year, datetime.date.today().month, 1)
 # verify that a new file needs to be downloaded
 verif_check = date_check(file_date=firstDayOfMonth, cursor=cursor)
-if verif_check:
-    logger.info('file exists, pass')
-    pipeline_title = 'No Companies House File'
-    pipeline_message = f'Pipeline closed for today'
-    pipeline_hexcolour = '#8f0d1a'
-    pipeline_messenger(title=pipeline_title, text=pipeline_message, hexcolour=pipeline_hexcolour)
-    quit()
+# if verif_check:
+#     logger.info('file exists, pass')
+#     pipeline_title = 'No Companies House File'
+#     pipeline_message = f'Pipeline closed for today'
+#     pipeline_hexcolour = '#8f0d1a'
+#     pipeline_messenger(title=pipeline_title, text=pipeline_message, hexcolour=pipeline_hexcolour)
+#     quit()
+#
+if len(os.listdir('file_downloader/files/fragments')) == 1:
+    ch_file, ch_upload_date = collect_companieshouse_file(firstDayOfMonth)
+    str_ch_file = str(ch_file)
+    logger.info('unzipping file')
+    unzipped_ch_file = unzip_ch_file(ch_file)
+    fragment_ch_file(f'file_downloader/files/{unzipped_ch_file}')
+    os.remove(f'file_downloader/files/{unzipped_ch_file}')
 
-ch_file, ch_upload_date = collect_companieshouse_file(firstDayOfMonth)
-str_ch_file = str(ch_file)
-logger.info('unzipping file')
-unzipped_ch_file = unzip_ch_file(ch_file)
-fragment_ch_file(f'file_downloader/files/{unzipped_ch_file}')
 fragment_list = os.listdir('file_downloader/files/fragments/')
-os.remove(f'file_downloader/files/{unzipped_ch_file}')
 for fragment in fragment_list:
     print(fragment)
     if fragment != 'fragments.txt':
         logger.info(fragment)
         st = time.time()
-        parse_fragment(f'file_downloader/files/fragments/{fragment}', host=host, user=user, passwd=passwd, db=database)
-        load_fragment(cursor, db)
+        parse_fragment(f'file_downloader/files/fragments/{fragment}', host=host, user=user, passwd=passwd, db=database, cursor=cursor, cursordb=db)
         logger.info('------')
         os.remove(f'file_downloader/files/fragments/{fragment}')
         et = time.time()
@@ -62,7 +69,8 @@ for fragment in fragment_list:
         logger.info(f'parse time for this iteration: {final_time}')
     else:
         pass
-
+upsert_sic(cursor=cursor, db=db)
+upsert_address(cursor=cursor, db=db)
 # when done, update filetracker
 filetracker_tup = (str_ch_file, ch_upload_date, datetime.datetime.now(), datetime.datetime.now())
 cursor.execute("""insert into BasicCompanyData_filetracker (filename, ch_upload_date, lastDownloaded, lastProcessed) VALUES (%s, %s, %s, %s)""", filetracker_tup)
